app/requirements.txt
Flask==3.0.0
setuptools==65.5.1
wheel==0.38.1
PyJWT==2.8.0
python-dotenv==0.19.2
